+++ src/components/temp-landing/AnimatedBackground.jsx
