+++ src/components/temp-landing/CodeSnippet.jsx
